fix failing test+move ViewExt
