+++ .github/pull_request_template.md
@@ -1,8 +1,11 @@
 ## Details
+
 * What did you do?
 
 ## Checklists
+
 ### Testing
+
 * [ ] New and existing unit tests pass
 * [ ] Unit tests have been written for all new computation
   * I __don't care__ how much work it is. Write the tests. Your future self will thank you. In this repo, we ♥ tests.
@@ -12,17 +15,20 @@
 * [ ] Espresso tests have been written for all UI changes
 
 ### Images
+
 * [ ] Any new/deleted images have been updated in
   * [ ] the image attributions fragment
   * [ ] the README
 
 ### Documentation
-* [ ] Any new or changed settings have been updated as described in the settings README
-  * This includes documentation in the main README
+
+* [ ] Any new or changed settings have been updated in the README
 * [ ] If function signatures changed, the docstrings have been updated
 
 ### Other
+
 * [ ] ktlint has run successfully
 
 ## Comments
+
 Fun but optional
