+++ app/src/main/kotlin/xyz/lbres/trickcalculator/ext/view/ViewExt.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.trickcalculator.utils
+package xyz.lbres.trickcalculator.ext.view
 
 import android.view.View
